feat: refactor certificates management with new components for better structure and maintainability
