+++ src/components/dashboard-student/certificates/index.ts
@@ -1,0 +1,4 @@
+export { default as CertificatesClient } from "./CertificatesClient";
+export { default as CertificatesHeader } from "./CertificatesHeader";
+export { default as CertificatesTable } from "./CertificatesTable";
+export { default as CertificateRow } from "./CertificateRow";
